backend/pyproject.toml
[tool.poetry]
name = "commodities-compass"
version = "1.0.0"
description = "Business Intelligence application for commodities trading"
authors = ["Your Name <email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
# Core
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
python-multipart = "^0.0.9"

# Database
sqlalchemy = "^2.0.27"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"

# Data Processing
pandas = "^2.2.0"
numpy = "^1.26.4"
openpyxl = "^3.1.2"  # For Excel file reading

# Authentication
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-decouple = "^3.8"
authlib = "^1.3.0"

# API & Validation
pydantic = "^2.6.1"
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0"

# Google Sheets API
google-api-python-client = "^2.118.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.0"

# Background Tasks (for future use)
celery = "^5.3.4"
redis = "^5.0.1"

# Utilities
httpx = "^0.26.0"
python-dateutil = "^2.8.2"
pytz = "^2024.1"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.4"
ruff = "^0.3.5"
pyright = "^1.1.350"
pre-commit = "^3.6.0"

[tool.poetry.scripts]
dev = "app.main:main"
lint = "app.lint:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"
asyncio_mode = "auto"

[tool.isort]
profile = "black"
line_length = 88
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
